+++ data/duke.txt
@@ -1,2 +1,5 @@
-E | 1 | work | 2012-12-21T18:22
-D | 0 | return book | 2010-08-15T09:23
+D | 1 | return book | 2010-08-15T09:23
+D | 1 | return book | 2010-08-15T09:25
+E | 1 | return horhd | 2013-08-15T15:34
+T | 1 | work
+T | 0 | hi
